# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Pourier API'
  description: "API REST pour la plateforme de vente de photos Pourier. Cette API permet aux utilisateurs d'acheter des photos, aux photographes de vendre leurs créations, et aux administrateurs de gérer la plateforme."
  contact:
    name: 'Pourier Support'
    email: contact@pourier.com
  license:
    name: Proprietary
    url: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'Serveur API Pourier'
paths:
...
  /api/photographer/revenue:
    get:
      tags:
        - 'Photographer - Revenue'
      summary: 'List all revenue records'
      description: 'Get paginated list of all revenue records for the photographer. Each sale generates revenue after 20% platform commission.'
      operationId: getPhotographerRevenue
      responses:
        200:
          description: 'Revenue records retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      data:
                        type: array
                        items:
                          properties:
                            id:
                              type: string
                              format: uuid
                            photographer_id:
                              type: string
                              format: uuid
                            photo_id:
                              type: string
                              format: uuid
                            photo:
                              properties:
                                id:
                                  type: string
                                  format: uuid
                                title:
                                  type: string
                                  example: 'Sunset over Ouagadougou'
                              type: object
                            photographer_amount:
                              description: "Photographer's share (80%) in FCFA"
                              type: number
                              format: float
                              example: 4000
                            platform_commission:
                              description: 'Platform commission (20%) in FCFA'
                              type: number
                              format: float
                              example: 1000
                            sold_at:
                              type: string
                              format: date-time
                              example: '2025-11-10T14:30:00Z'
                            available_at:
                              description: 'When revenue becomes available for withdrawal'
                              type: string
                              format: date-time
                              example: '2025-11-20T14:30:00Z'
                          type: object
                      current_page:
                        type: integer
                        example: 1
                      per_page:
                        type: integer
                        example: 20
                      total:
                        type: integer
                        example: 80
                    type: object
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
      security:
        -
          bearerAuth: []
  /api/photographer/revenue/available:
    get:
      tags:
        - 'Photographer - Revenue'
      summary: 'Get available balance'
      description: "Get the total available balance ready for withdrawal (revenue where available_at <= now). This is the photographer's 80% share after 20% platform commission."
      operationId: getPhotographerRevenueAvailable
      responses:
        200:
          description: 'Available balance retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      available_amount:
                        description: 'Total available balance in FCFA ready for withdrawal'
                        type: number
                        format: float
                        example: 50000
                    type: object
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
      security:
        -
          bearerAuth: []
  /api/photographer/revenue/pending:
    get:
      tags:
        - 'Photographer - Revenue'
      summary: 'Get pending revenue'
      description: 'Get all revenue records that are not yet available for withdrawal (available_at > now). Typically held for a period before becoming available.'
      operationId: getPhotographerRevenuePending
      responses:
        200:
          description: 'Pending revenue retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: string
                          format: uuid
                        photographer_id:
                          type: string
                          format: uuid
                        photo_id:
                          type: string
                          format: uuid
                        photo:
                          properties:
                            id:
                              type: string
                              format: uuid
                            title:
                              type: string
                              example: 'Mountain Landscape'
                          type: object
                        photographer_amount:
                          description: "Photographer's share (80%) in FCFA"
                          type: number
                          format: float
                          example: 4000
                        sold_at:
                          type: string
                          format: date-time
                          example: '2025-11-12T10:00:00Z'
                        available_at:
                          description: 'When this revenue will become available'
                          type: string
                          format: date-time
                          example: '2025-11-22T10:00:00Z'
                      type: object
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
      security:
        -
          bearerAuth: []
  /api/photographer/revenue/history:
    get:
      tags:
        - 'Photographer - Revenue'
      summary: 'Get revenue history by date'
      description: "Get daily revenue summary with total earnings and sales count, grouped by date. Shows photographer's 80% share after 20% platform commission."
      operationId: getPhotographerRevenueHistory
      responses:
        200:
          description: 'Revenue history retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      data:
                        type: array
                        items:
                          properties:
                            date:
                              description: 'Date of sales'
                              type: string
                              format: date
                              example: '2025-11-14'
                            total:
                              description: 'Total photographer earnings for the day in FCFA'
                              type: number
                              format: float
                              example: 12000
                            sales:
                              description: 'Number of sales on this date'
                              type: integer
                              example: 3
                          type: object
                      current_page:
                        type: integer
                        example: 1
                      per_page:
                        type: integer
                        example: 30
                      total:
                        description: 'Total number of days with sales'
                        type: integer
                        example: 90
                    type: object
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    UnauthorizedResponse:
      title: 'Unauthorized Response'
      description: 'Réponse pour une authentification requise'
      properties:
        message:
          type: string
          example: 'Non authentifié'
      type: object
...
  responses:
    UnauthorizedResponse:
      description: 'Unauthorized - Authentication required'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Enter JWT Bearer token obtained from /api/auth/login endpoint'
...
security:
  -
    bearerAuth: []